# shared/article/api.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: Example.com
servers:
  - url: "/api"
paths:
  "/article/{number}":
    get:
      security: []
      summary: "Query a article"
      operationId: "getArticle"
      parameters:
        - name: number
          in: path
          description: the number of the article
          schema:
            type: string
      responses:
        200:
          description: the article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        404:
          description: the Article was not found
  "/article/{number}/bom":
    get:
      security: []
      summary: get a bom of the article
      operationId: "getBom"
      parameters:
        - name: number
          in: path
          description: the number of the article
          schema:
            type: string
      responses:
        200:
          description: the bom of the article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bom"
        404:
          description: the Article was not found
  "/article/{number}/file":
    post:
      security: []
      summary: uploads new files to the specified article
      operationId: "uploadFiles"
      parameters:
        - name: number
          in: path
          description: the number of the article the file belongs to
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                attachments:
                  type: array
                  items:
                    type: string
                    format: binary
    get:
      security: []
      summary: get the default files for the article
      operationId: "getArticleFiles"
      parameters:
        - name: number
          in: path
          schema:
            type: string
      responses:
        200:
          description: the default files of the article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        404:
          description: the article or file was not found
  "/article/{number}/file/{path}":
    get:
      security: []
      summary: get the specified file
      operationId: "getFile"
      parameters:
        - name: number
          in: path
          schema:
            type: string
        - name: path
          in: path
          schema:
            type: string
      responses:
        200:
          description: the specified file
          headers:
            cache-control:
              description: the max length this content is valid
              schema:
                type: string
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        404:
          description: the article or file was not found
components:
  schemas:
    Article:
      type: object
      properties:
        number:
          description: Article number
          type: string
          minLength: 7
          example: 9000004
        files:
          description: The folder structure found in the articles folder
          type: array
          items:
            $ref: "#/components/schemas/FileSystem"
          minLength: 1

    FileSystem:
      type: object
      properties:
        name:
          type: string
          description: the current folder name
        path:
          type: string
          description: the path to the folder
        type:
          type: string
          enum:
            - "file"
            - "folder"
        children:
          type: array
          items:
            $ref: "#/components/schemas/FileSystem"
    Bom:
      type: object
      properties:
        index:
          type: string
          description: the index of the article
        number:
          type: string
          description: the article number
        name:
          type: string
          description: the name of the article
        children:
          type: array
          items:
            $ref: "#/components/schemas/Bom"